requirements.txt
openai>=1.30.0
tqdm>=4.66.0
python-dotenv>=1.0.0